add petdetail data, view list petdetail, add to cart, view cart

+++ backend/src/main/java/com/dut/backend/repository/PetRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PetRepository  extends JpaRepository<Pet, Integer> {
+public interface PetRepository  extends JpaRepository<Pet, Long> {
     @Query("SELECT  p.breed FROM Pet p ")
     List<String> findBreeds();
     Pet findByBreed(String breed);
